采购入库，分批输入入库数量

+++ erp_entity/src/main/java/cn/itcast/erp/entity/Storedetail.java
@@ -7,7 +7,7 @@ public class Storedetail {
 	private Long uuid;//编号
 	private Long storeuuid;//仓库编号
 	private Long goodsuuid;//商品编号
-	private Long num;//数量
+	private Double num;//数量
 
 	public Long getUuid() {		
 		return uuid;
@@ -27,10 +27,10 @@ public class Storedetail {
 	public void setGoodsuuid(Long goodsuuid) {
 		this.goodsuuid = goodsuuid;
 	}
-	public Long getNum() {
+	public Double getNum() {
 		return num;
 	}
-	public void setNum(Long num) {
+	public void setNum(Double num) {
 		this.num = num;
 	}
